requirements.txt
# Core Dependencies
MetaTrader5>=5.0.45
numpy>=1.21.0
pandas>=1.5.0

# GUI
tk>=0.1.0

# Telegram Bot
python-telegram-bot>=20.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.6.0
joblib>=1.1.0

# Performance Optimization
numba>=0.56.0

# System Monitoring
psutil>=5.9.0
GPUtil>=1.4.0

# Timezone Support
pytz>=2021.3

# Data Visualization (Optional - lazy loaded)
matplotlib>=3.5.0

# Excel & PDF Conversion
openpyxl>=3.0.0
fpdf2>=2.7.0
cryptography>=41.0.0
matplotlib>=3.5.0

# Testing (Optional)
pytest>=7.0.0
nuitka